reorganize train folder
