+++ clinicadl/random_search/random_search_utils.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, Tuple
 
 import toml
 
-from clinicadl.train.tasks.base_training_config import Task
-from clinicadl.train.train_utils import extract_config_from_toml_file
+from clinicadl.train.tasks import Task
+from clinicadl.train.utils import extract_config_from_toml_file
 from clinicadl.utils.exceptions import ClinicaDLConfigurationError
 from clinicadl.utils.preprocessing import path_decoder, read_preprocessing
