--- clinicadl/train/__init__.py
@@ -1,2 +1,0 @@
-from .tasks.base_training_config import BaseTaskConfig
-from .train_utils import preprocessing_json_reader
