+++ clinicadl/train/from_json/__init__.py
@@ -1,0 +1,1 @@
+from .from_json_cli import cli
