+++ clinicadl/train/list_models/__init__.py
@@ -1,0 +1,1 @@
+from .list_models_cli import cli
