+++ clinicadl/train/list_models/list_models_cli.py
@@ -28,10 +28,6 @@ def cli(
     model_layers,
 ):
     """Show the list of available models in ClinicaDL."""
-    from .train_utils import get_model_list
+    from clinicadl.train.utils import get_model_list
 
     get_model_list(architecture, input_size, model_layers)
-
-
-if __name__ == "__main__":
-    cli()
