+++ clinicadl/train/resume/__init__.py
@@ -1,0 +1,1 @@
+from .resume_cli import cli
