+++ clinicadl/train/resume/resume_cli.py
@@ -1,12 +1,17 @@
 import click
 
 from clinicadl.utils import cli_param
-from clinicadl.utils.cli_param import train_option
 
 
 @click.command(name="resume", no_args_is_help=True)
 @cli_param.argument.input_maps
-@train_option.split
+@cli_param.option_group.cross_validation.option(
+    "--split",
+    "-s",
+    type=int,
+    multiple=True,
+    help="Train the list of given splits. By default, all the splits are trained.",
+)
 def cli(input_maps_directory, split):
     """Resume training job in specified maps.
 
@@ -15,7 +20,3 @@ def cli(input_maps_directory, split):
     from .resume import automatic_resume
 
     automatic_resume(input_maps_directory, user_split_list=split)
-
-
-if __name__ == "__main__":
-    cli()
