+++ clinicadl/train/tasks/__init__.py
@@ -1,0 +1,1 @@
+from .base_task_config import BaseTaskConfig, Task
