+++ clinicadl/train/tasks/classification/__init__.py
@@ -1,0 +1,2 @@
+from .classification_cli import cli
+from .classification_config import ClassificationConfig
