+++ clinicadl/train/tasks/regression/__init__.py
@@ -1,0 +1,2 @@
+from .regression_cli import cli
+from .regression_config import RegressionConfig
