+++ clinicadl/train/utils.py
@@ -6,12 +6,12 @@ import click
 import toml
 from click.core import ParameterSource
 
-from clinicadl.train import BaseTaskConfig
-from clinicadl.train.tasks.base_training_config import Task
 from clinicadl.utils.caps_dataset.data import CapsDataset
 from clinicadl.utils.exceptions import ClinicaDLConfigurationError
 from clinicadl.utils.maps_manager.maps_manager_utils import remove_unused_tasks
 from clinicadl.utils.preprocessing import path_decoder, read_preprocessing
+
+from .tasks import BaseTaskConfig, Task
 
 
 def extract_config_from_toml_file(config_file: Path, task: Task) -> Dict[str, Any]:
